[mkcal] Don't delete custom properties on incidence deletion. Contributes to MER#916
In order to support sync, we need to store data in CUSTOMPROPERTIES
for incidences: usually ETAG and URI values, which are not supported
as COMPONENT values.

Previously, mkcal would delete all values in the CUSTOMPROPERTIES
table on incidence deletion - but this would make it difficult to
support upsyncing local incidence deletions due to the data loss.

This commit ensures that CUSTOMPROPERTIES data is retained after
incidence deletion.

Contributes to JB#916