Support optional notebook for some service handler interface methods
Just having storage and calendar parameters relied on passed event
having been added to some notebook, and that way knowing which account
to use for sending updates.

That is not, however, very practical when removing attendees from an
event. At the same the event should be canceled with removed attendees,
and should be in normal state with the existing attendees.