[mkcal] Read floating_date dates from no time-zone column.
Previously floating_date dates where read using the UTC column.
This is creating issues when saving and loading are not in the
same time zone sign. For instance, writing a date in a UTC-xx
time zone will put the number of seconds for the day after since
UTC is already the day after when the time zone is in -xx. Reading
from another -xx time zone is not an issue since the additional
day will be removed. But this is creating issues when reading
from a +xx time zone.