[mkcal] Fix some issues related to ClockTime handling. Contributes to JB#43708
When expanding events (to determine whether an event or its occurrences
should be returned within a given date-time range) we need to handle
both clock-time ranges, and clock-time events, specially.

That is, if an event is clock-time 10 am until clock-time 11 am, then
that event should be returned for any range covering 10 am -> 11 am,
regardless of what timezone is specified for that range.