[mkcal] Include TZ offset in Reminder (Alarm) datetimes. Contributes to JB#52684
Due to QTBUG-26161 it is not enough to call
  dateTime.toString(Qt::ISODate)
as this does not retain timezone offset information.

Instead, we have to first convert the datetime to being in
Qt::OffsetFromUTC spec before calling toString(Qt::ISODate).

Note: we DON'T apply the same transformation to the recurrenceId,
as that might cause issues elsewhere (e.g. when looking up the
occurrence from the database).