Port to CMake
This project uses KDE's KCalendarCore. However, that package can only be
used with build systems that support CMake dependencies, which QMake is
not one of. KCalendarCore probably shipped a pkgconfig file in the past,
but this has not been the case for a quite a while as far as I can see.

Since using QMake thus prevents using a current KCalendarCore, switch
the build system to CMake instead