[mkcal] Fix corrupted attachment info saved to the database. Fixes JB#57634
SQLITE_STATIC on bind_text promises the data doesn't need to be copied,
but here the QByteArray got out of scope before the sql query got
executed. Literally the docs assume validity until the statement
is finalized but until sqlite3_step() is what the other parts of the code
use and what could be assumed the end of interest to these.