[mkcal] Add a new storage observer signal on DB update
The existing storageModified() signal is emitted whenever
the ".changed" file is touched. Restrict its emission
only if some change has been done to the database content
from an external process. Currently emit it also if a
change is done in process to any notebook.

Add a new signal storageUpdated() that is emitted on any
modification of the database component parts (addition,
modification of deletion) to be in part with the associated
mCalendar.