[mkcal] Fix alarms for recurring events with exceptions.
Since the parent of a recurring event does not
contain anymore the EXDATE of the exceptions, the
getNextDateTime() in setAlarms() is always
returning the next time in the recurrence without
taking into account the exceptions. This makes
the code for recurring events in general broken
regarding the alarms, including cases like:
* change time of the next occurrence : alarm will
  ring twice, one time on the normal time and one
  time for the new time,
* delete the next occurrence : alarm will ring
  anyway,
* cancel one occurrence, alarm will ring anyway.

Add a new test to check alarms for recurring
events in the general case and the above ones.