[mkcal] Fix issue on recurring alarms.
Commit f295c422 introduced a fix for recurring alarms
with exceptions. The fix was written in the timed
plugin code to use the full recurring incidence
with exceptions to compute the next alarm occurrence.
This code is unit tested, but the place where it is
called (the mkcaltool helper) was not updated to
properly load the full incidence description,
including exceptions.