[mkcal] Adjust saving and loading attendees. JB#55317
Mkcal was having its own restrictions on what the stored attendee
content could be. Including that organizer wasn't allowed
in the attendee list of an incidence, the email addresses needed to be
unique among the participants and all attendees needed an email address.

Adjusted the system so that organizer and the attendees are saved as the
incidence has them. Only empty ones (=no email nor name) are filtered out.

As side-effect this is now the first (finally!) mechanism to update the
database version, and to have database version info to begin with.