[mkcal] Drop ExtendedStorage from service API.
Rationalise the ServiceHandler API, not passing the
ExtendedStorage pointer when it's not used in the actual
interface and only there to check notebook validity
with respect to the storage. Instead, always pass directly
the notebook the incidence is associated with. Let the
client retrieve this notebook, depending how the
incidence is stored.