[mkcal] Try to make DB migration less fatal on errors.
Commit changes immediately after migration to the next
version and bump user_version as soon as possible to
avoid the database being migrated but the pragma not
being in sync with the DB status.

Don't fail on migration error when adding a new column,
assuming that the column may already exists.