[mkcal] Fix discrepency between migrated and new DBs. Contributes to JB#59982
The INSERT and UPDATE sharing the same code in
sqliteformat.cpp, and INSERT relying on column
order, new column should be added at the end in
the CREATE statement. Otherwise, there is a
difference between migrated DBs where the new
column is at the end and new DBs where the column
has been added in the middle of other columns.

Correct this issue for the new thisAndFuture column.