[mkcal] Remove timezone from database storage.
The storage of timezones to check incidences upon
when the incidence timezone name is not matching
any of the system, is deprecated. KCalendarCore
is not providing anymore a timezone definition when
parsing ICS data. It is always using the system
timezone definitions to match any received ICS
data. There is thus no need anymore to store a
timezone defintion in the database.

Not to complexify the code, the existing
timezones table is kept as it is, simply not
used anymore.