[mkcal] Don't reload event by UID if already in memory.
Don't reload an existing incidence from DB.
Either the calendar is already in sync with
the calendar or the database has been externally
modified and in that case, the calendar has been
emptied.

Testing the parent incidence in case of recurring
incidence should be enough to guarantee the
existence of all exceptions since the load
methods ensure loading of whole series.