[mkcal] Order recurring events first when searching.
The search statement is sorted by date and only
a limited number of entries may be converted into
incidences. Since occurrences of matching
recurring events may happen at a later date than
the start date, order the recurring events always
first in the return list.

This allows for instance to get yearly recurring
events like birthday to still be in the return
list even when there more than the limit of
events before the birth date.