[mkcal] Purge incidence from given notebook.
Adding a notebook UID to the purgeDeletedIncidences() API
would allow to selectively remove incidences from the
database. Without it, it is not possible to leave
incidences marked as deleted in a notebook different
than the one we want to insert an incidence to with
the same UID (like when changing notebook for instance).